Small improvements, stuck
Fuck, realized a huge issue with the way I do my destruction system. Basically with the way I do it currently its impossible to fix the last couple issues.
So I will need to rewrite the impact solver and destruction solver.

Basically I need to know how the object will deform+break immediately when the impact occure. Currently I start a background thread and dont know the result until right before the next physics tick.
The worse part is that I knew about this problem all the time but ignored it and thought, I fix it later by faking it. And I have spent the last few days trying to fake it but its physics stuff so its easier said that done