Fixed issue with memory leak and bad pointer
Feels like im getting close to the limit of how stable I can get this without getting way too complicated.

Considering that the input I get from unity aint 100% reliable. For example if an object is outside then deep inside collider next frame the normals will point in all directions. When decollinding the velocity will be in the opposite direction, meaning it can report collision with the object its decolliding from, and if it does the velocity will be wrong.